refactor: reorganize new_user_activity module into subdirectory
Move all new_user_activity files into lib/new_user_activity subdirectory
to improve code organization and isolation. This ensures the auto_route
generator correctly scopes route generation and prevents mixing of
unrelated router definitions.

Changes:
- Move new_user_activity_bloc.dart, _page.dart, _router.dart, _screen.dart,
  _state.dart to lib/new_user_activity/ subdirectory
- Update internal imports to use relative paths
- Update router config to use generateForDir: ['lib/new_user_activity']
- Verified iOS and Android builds complete successfully
