+++ ui/user_activity/lib/new_user_activity/new_user_activity_bloc.dart
@@ -5,7 +5,7 @@ import 'package:core/core.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:movies/movies.dart';
 import 'package:user_activities_domain/domain.dart';
-import 'package:user_activity/new_user_activity_state.dart';
+import 'new_user_activity_state.dart';
 
 class NewUserActivityCubit extends Cubit<NewUserActivityState> {
   final SearchMovies _searchMovies;
