+++ ui/user_activity/lib/new_user_activity/new_user_activity_page.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
 import 'package:movies/movies.dart';
 import 'package:user_activities_domain/domain.dart';
-import 'package:user_activity/new_user_activity_bloc.dart';
-import 'package:user_activity/new_user_activity_screen.dart';
+import 'new_user_activity_bloc.dart';
+import 'new_user_activity_screen.dart';
 
 @RoutePage()
 class NewUserActivityPage extends StatefulWidget {
