+++ ui/user_activity/lib/new_user_activity/new_user_activity_router.dart
@@ -1,9 +1,9 @@
 import 'package:auto_route/auto_route.dart';
-import 'package:user_activity/new_user_activity_page.dart';
+import 'package:user_activity/new_user_activity/new_user_activity_page.dart';
 
 part 'new_user_activity_router.gr.dart';
 
-@AutoRouterConfig(generateForDir: ['lib'])
+@AutoRouterConfig(generateForDir: ['lib/new_user_activity'])
 class NewUserActivityRouter extends RootStackRouter {
   @override
   List<AutoRoute> get routes => [
